shopping cart with items

+++ backend/src/resolvers/Mutation.js
@@ -3,12 +3,24 @@ const jwt = require('jsonwebtoken')
 const { randomBytes } = require('crypto')
 const { promisify } = require('util')
 const { transport, makeANiceEmail } = require('../mail')
+const { hasPermission } = require('../utils')
 
 const Mutations = {
 
   async createItem(parent, args, ctx, info) {
+    if ( ! ctx.request.userId) {
+      throw new Error(`You must be logged in to do that!`)
+    }
     const item = await ctx.db.mutation.createItem({
-      data: { ...args }
+      data: {
+        // create a relationship between Item and User
+        user: {
+          connect: {
+            id: ctx.request.userId,
+          },
+        },
+        ...args,
+      }
     }, info)
     return item
   },
@@ -28,9 +40,13 @@ const Mutations = {
   async deleteItem(parent, args, ctx, info) {
     const where = { id: args.id }
     // 1. find the item
-    const item = await ctx.db.query.item({ where }, `{ id title }`)
+    const item = await ctx.db.query.item({ where }, `{ id title user { id } }`)
     // 2. check permissions
-    // TODO
+    const ownsItem = item.user.id === ctx.request.userId
+    const hasPermissions = ctx.request.user.permissions.some(permission => ['ADMIN','ITEMDELETE'].includes(permission))
+    if ( ! ownsItem && ! hasPermission) {
+      throw new Error('Not allowed')
+    }
     // 3. delete it
     return ctx.db.mutation.deleteItem({ where }, info)
   },
@@ -151,7 +167,82 @@ const Mutations = {
     })
     // return new user
     return updatedUser
-  }
+  },
+
+  async updatePermissions(parent, args, ctx, info) {
+    // check if logged in
+    if ( ! ctx.request.userId) {
+      throw new Error(`You must be logged in to do that!`)
+    }
+    // query current user
+    const currentUser = await ctx.db.query.user({ 
+      where: { id: ctx.request.userId }
+    }, info)
+    // check permissions to do this
+    hasPermission(currentUser, ['ADMIN', 'PERMISSIONUPDATE'])
+    // update permissions
+    return await ctx.db.mutation.updateUser({
+      data: {
+        permissions: { set: args.permissions },
+      },
+      where: {
+        id: args.userId,
+      }
+    },
+    info)
+  },
+
+  async addToCart(parent, args, ctx, info) {
+    // check if logged in
+    const userId = ctx.request.userId
+    if ( ! userId) {
+      throw new Error(`You must be logged in to do that!`)
+    }
+    // query users cart
+    const [existingCartItem] = await ctx.db.query.cartItems({
+      where: {
+        user: { id: userId },
+        item: { id: args.id },
+      }
+    })
+    // check if item already in cart
+    if (existingCartItem) {
+      // increment 
+      return ctx.db.mutation.updateCartItem({
+        where: { id: existingCartItem.id },
+        data: { quantity: existingCartItem.quantity+1 }
+      }, info)
+    }
+    // create fresh CartItem
+    return ctx.db.mutation.createCartItem({
+      data: {
+        user: {
+          connect: { id: userId },
+        },
+        item: {
+          connect: { id: args.id },
+        },
+        quantity: 1,
+      }
+    }, info)
+  },
+
+  async removeFromCart(parent, args, ctx, info) {
+    // find the CartItem
+    const cartItem = await ctx.db.query.cartItem({ where: { id: args.id }}, '{ id, user { id }}')
+    // make sure we found an item
+    if ( ! cartItem) {
+      throw new Error('No cart item found')
+    }
+    // make sure they own that cart item
+    if (cartItem.user.id !== ctx.request.userId) {
+      throw new Error('Cheatin huhhh')
+    }
+    // delete cartItem
+    return ctx.db.mutation.deleteCartItem({
+      where: { id: args.id },
+    }, info)
+  },
 }
 
-module.exports = Mutations;
+module.exports = Mutations
